Fix reporting certain config syntax errors
Change I0421c5a446d0b75194096b1d8f0e0866dae3b8f0 began including
BranchMatcher objects instead of strings in the SourceContext class
if an implied-branches pragma was supplied.  If a syntax error
occurs in a repo with such a pragma, the source context needs to
be serialized to zk, and that would fail.  To correct this, add
serialization methods to the source context and branch matchers.

Also correct an unrelated error where the syntax "- job:" would
fail to report the error to the user because the value None was
added to the list of config objects, and we always assume that
is a dict.  To correct this, raise the syntax error earlier in
that case.

Change-Id: Id7c1dff5e9865d21c3215bdc32c8fa2cb6603aaa