Fix updating change when emulating gerrit topic support
We conditionally update some change attributes within the pipeline
manager.  One of those is the topic_needs_change attribute, however,
we wrap updating that by checking whether a different attribute
(commit_needs_change) attribute.

This meant that if a new patchset was uploaded, other changes which
had previously seen the old patchset would not necessarily be updated
to see the new one.

Since setChangeAttributes now performs its own internal check to
see if an update is necessary, we can simply remove the conditional
so that it always runs.

Change-Id: I9662e2d230d88dcd19e02f2def2385965fc386a8