Update patch set 1
Patch Set 1:

I don't dispute that at least one person misunderstood this, however, I don't think we should start including a lot of documentation for things that "aren't".  There are a lot of ways people might misinterpret things, and it's only going to get more confusing if we list them all.

I'm happy to entertain suggestions about how to improve the documentation about things that "are".

Perhaps the issue arose because the reader read the first paragraph of the branches settings here, without having read the introduction to the job section above where variants are described.  Perhaps a better approach would be to highlight that again here, or link to it.

Patch-set: 1