Bump up to python3.10
This adds python3.10 testing on Jammy and switches the docker images to
python3.10 from 3.8.

We run sudo for postgres with -Hi to avoid non fatal errors when
postres' client attempts to write command history to Zuul's homedir (it
is running as the postgres user which can't write to zuul's homedir). We
also need to update the libffi package version for jammy to 8 in
bindep.txt. Finally, python_version values need to be quoted as "3.10"
is different than 3.10 which is equivalent to 3.1 when serialized by
yaml as a float.

Update the ansible management tooling to better detect the bin path in
virtualenvs. It seems that on ubuntu-jammy virtualenv will create a venv
with venv/local/bin as the bin path instead of venv/bin. I do not know
if this is an Ubuntu behavior or python3.10 behavior. On OpenSUSE
Tumbleweed with python3.10 virtualenv produces venv/bin.

Finally we switch the bulk of jobs over to using nodeset: ubuntu-jammy
as the default python there is 3.10.

Change-Id: I97b90bb7a23c90f108f23dda9fdd0e89f9f4dbca