Add feature to exclude jobs based on branch
A new parameter called 'include-branch-pattern' can be used to restrict
which job variants are considered valid based on the branch on which
they are defined.

This is useful for the use case where jobs may be defined on a given
branch that new branches may be forked from, but those new branches only
want a subset of the parent job definitions by default.

Change-Id: Iee7d562b60e9fdb501172d2b0e50cfb16030d2b3
Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>