Add USER to Dockerfile
We create a zuul user in the image, but we don't switch to it in
the dockerfile, which leaves it up to users to ensure that they
specify a user; otherwise the containers will run as root by default
which is not recommended.

An exception is made for zuul-fingergw since it is typically
configured to drop privileges.

Change-Id: I05b9cfd1cc8427384d1887c767e9751299e797d2