Set logfile isFetching initial state to true
Setting the logfile isFetching initial state to true is more accurate to
reality as when the logfile object is first insantiated it has not been
fetched and is going to be fetched. This is important to ensure our
useEffect() callbacks file and apply at the right times otherwise they
will update values too early, fire, then not fire again when the
fetching is complete. If they fire before fetching is complete then they
do not function.

Change-Id: Ic8cd2e4ab2d2d7fd5f74ff6862f719c0aaa756dc