Update patch set 1
Patch Set 1:

I do not believe it was the intention that the container images would support running zuul-manage-ansible within the container themselves; nor do I think they should.  In general, container images should be considered to be immutable, and the reason that we run zuul-manage-ansible in our Dockerfile is explicitly to freeze that output into the container image.

I think if a user wants to add to the set of software installed on the image, then that should be done by updating the image, not altering the container.

So a user would have a Dockefile with: FROM this image; USER root; run z-m-a; USER zuul.  Then there's a new image with the requested software.

Patch-set: 1
Attention: {"person_ident":"Gerrit User 1 \u003c1@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"REMOVE","reason":"removed on reply"}