Add always-dynamic-branches option
This adds an option to specify that certain branches should always trigger
dynamic configuration and never be included in static configuration.

The use case is a large number of rarely used feature branches, where
developers would still like to be able to run pre-merge check jobs, and alter
those jobs on request, but otherwise not have the configuration clogged up
with hundreds of generally unused job variants.

Change-Id: I60ed7a572d66a20a2ee014f72da3cb7132a550da