Don't clear branch cache on reconfig
Clearing the branch cache for reconfigurations can break zuul-web which
relies on the branch cache in Zookeeper since it can't fetch project
branches directly.

The branch cache will also be cleared out when issuing a tenant
reconfigure command which goes through the same code path. In this case
the branch cache will only be filled gradually.

Change-Id: I8560f245a6da0e7a48ee04ac6c6adaa36b1f26b8