Update patch set 1
Patch Set 1:

We similarly discovered that network interface interrogation in the Ansible stdlib relies on the netaddr package, but doesn't declare a dependency on it (hence why it got appended to the list recently). Dropping ARA caused a regression for a bunch of jobs in OpenDev (and likely elsewhere) because ARA declared a dependency on netaddr and so was implicitly providing it in the venv.

This raises a larger question, should Ansible be delcaring dependencies on modules used by its stdlib?

Patch-set: 1
CC: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>