Fix zuul-web layout update on full/tenant-reconfigure
When an operator issues a full-reconfigure or tenant-reconfigure
command to the scheduler, the branch cache is cleared.  It will
be fully populated again during the subsequent full-reconfiguration,
or eventually populated more slowly over time after a
tenant-reconfiguration.  But during the time the branch cache is
clear, zuul-web will be unable to update layouts since it is incapable
of populating the branch cache itself.

This produces a short time period of errors during a full-reconfig
or a potentially long time period of errors after a tenant-reconfig.

To correct this, we now detect whether a layout update in zuul-web
is incomplete due to a branch cache error, and retry the update
later.  In the scheduler, we only clear the projects from the branch
cache that are affected by the tenants we are reloading (still all the
projects for a full-reconfigure).  This limits the time during which
zuul-web will be unable to update the layout to only the time that
the scheduler spends actually performing a reconfiguration.

(Note that in general, the error here is not because zuul-web is
loading the layout for the tenant that zuul-scheduler is reconfiguring,
but rather it is loading a layout for a tenant which has projects
in common with the tenant that is being reconfigured.)

Change-Id: I6794da4d2316f7df6ab302c74b3efb5df4ce461a