Update patch set 2
Patch Set 2:

Because the full-/tenant-reconfigure commands are mostly there to fix problems if something gets out of sync, I think it would be better to retain the cache clearing behavior (rather than add per-project ltimes).  Plus, it's a little unclear to me what ltime we would use for the projects (we use the write time of the cache itself for the cache min ltime, but that would be after the point where we fetched the branches, so it might allow us to get out of sync?  and if we get an ltime before we fetch the branches and use that, then it'll always be older than the cache ltime and we'll think we're out of date).

For these reasons, I propose an alternate approach in https://review.opendev.org/841818

Patch-set: 2
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Attention: {"person_ident":"Gerrit User 27582 \u003c27582@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}