Add global semaphore support
This adds support for global semaphores which can be used by multiple tenants.
This supports the use case where they represent real-world resources which
operate independentyl of Zuul tenants.

This implements and removes the spec describing the feature.  One change from
the spec is that the configuration object in the tenant config file is
"global-semaphore" rather than "semaphore".  This makes it easier to distinguish
them in documentation (facilitating easier cross-references and deep links),
and may also make it easier for users to understand that they have distinct
behavoirs.

Change-Id: I5f2225a700d8f9bef0399189017f23b3f4caad17