Don't fail config loading on missing min. ltime
If the list of branches for a project includes items that are not (yet)
in the min. ltimes for a layout state we can end up in a situation where
a scheduler is unable to start.

2022-05-16 17:34:50,895 ERROR zuul.Scheduler: Error starting Zuul:
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/cmd/scheduler.py", line 98, in run
    self.sched.prime(self.config)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 931, in prime
    tenant = loader.loadTenant(
  File "/opt/zuul/lib/python3.8/site-packages/zuul/configloader.py", line 2530, in loadTenant
    new_tenant = self.tenant_parser.fromYaml(
  File "/opt/zuul/lib/python3.8/site-packages/zuul/configloader.py", line 1631, in fromYaml
    self._cacheTenantYAML(abide, tenant, loading_errors, min_ltimes,
  File "/opt/zuul/lib/python3.8/site-packages/zuul/configloader.py", line 1896, in _cacheTenantYAML
    pb_ltime = min_ltimes[project.canonical_name][branch]
KeyError: 'new_branch'

The difference could be due to a missed branch creation event or a
simple race condition. The later case might fix itself after the
reconfig triggered by the branch creation event was processed.

Change-Id: I1838e66bc5296f153aa4c7a83ac0addb6c4db1aa