zuul-web: harmonize behavior for boolean parameters
Some boolean-like parameters in search queries were expected to be
set to 1 or 0; others to 'true' or anything else. Harmonize the
behavior: now every boolean-like parameter, if set, must be equal to
'true' or 1 if true; any other value will map to false.

Add the "Final" search criteria to the builds page in the web UI.

Change-Id: I78558d8e83df3fbb0cad61a5220cbaf7bf2c2023