Fix the model API 8 transition
Model API 8 adds the deduplcate attribute to FrozenJobs. It didn't
properly check the model api before processing the deduplicate attribute
during FrozenJob equality comparisons and FrozenJob serializations. Fix
this by ignoring the the attriute when the model API is less than 8.

An example of the errors this produced previously can be found at:
  https://paste.opendev.org/show/bHNqoDi2M2f9ExF9s1NH/

Change-Id: Id6168579b0c776213d489fe53fcb0835783eb781