Fix result updating in deduplication
We didn't correctly filter the list of items which contain deduplicated
builds and would therefore potentially call setResult on non-live items.

This change corrects the filter and error.

This error can be observed in the test output, however it a non-fatal
error and is therefore difficult to asses with our test infrastructure.

It can occasionally cause some tests to fail depending on race conditions.

Due to the infeasability, a test is not included.

Change-Id: I396b2e0e43b9e6ec501117d427e3c483962a2761