Dowgrade pipeline summary refresh error to warning
If there is no node in ZK for a pipeline summary, we will log an
exception and continue to use the old data, if any.  This happens
on system bootstrapping, and is harmless.  It could conceivably
happen during operation as well, but is more likely to show up as
a data corruption error, rather than NoNode.

Reduce the severity of the specific case of NoNodeError since it
is generally expected during system bootstrapping.  This logs it
as a warning with no traceback.

If it persists, it should still be obvious to operators, but it
no longer looks like an error that needs attention.

(Everything above is true for developers running unit tests as well.)

Change-Id: Ifb2a176cf79b4f536f7cf7c54afb8798af98b4ad