Temporarily pin OpenStackSDK before 0.99
The OpenStack SDK/CLI team made an experimental "release candidate"
with 0.99.0, and OpenDev observed missing CORS headers for objects
uploaded to RackSpace's Swift service after restarting executors
onto container images with this version installed. Pin to an earlier
version for now, in order to avoid this and any other as of yet
undiscovered problems in that version.

Change-Id: If1cf1f8c301de09df1d212b6cef151317f6dc6bf