Update patch set 2
Patch Set 2:

(1 comment)

Any reason not to be as generous on the API side as we are on the javascript side?  That is, accept 0, 1, '0', '1' as well?

If we did that, I think we could merge this without a release note.  Otherwise, it probably warrants a release note in case people have "=0" encoded in queries or automation.

Patch-set: 2
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>