Parallelize config cache loading
Loading config involves significant network operations for each project:

* Loading project keys
* Asking the source for the list of branches for each project
* Retrieving the config file contents from the ZK cache (if present)
* Retrieving the config file contents from git (otherwise)

Only the third item in that list is parallelized currently; the others
are serialized.  To parallelize the remainder, use a thread pool executor.

The value of max_workers=4 is chosen as it appears in practice on OpenDev
to make the most significant reduction in startup time while higher values
make little difference (and could potentially contribute to DoS scenarios
or local thread contention).  Observed config priming times for various
worker counts:

1: 282s
2: 181s
4: 144s
8: 146s

Change-Id: I65472a8af96ed95eb28b88cc623ef103be76a46f