Use attributes instead of a nested dict for web cache
This is a code-style only change which uses instance attributes
instead of a nested dict to store cache information in zuul-web.

This is slightly shorter, easier to read/type, less subject to
typos, and allows static analysis (eg pyflakes) to work better.

It also was slighly misleading to have the two caches combined
into a single variable since they operated very differently (one
is a single cache, the other is a set of caches).

Change-Id: I29ece7e8a39992724596d2d252ad7b023c50f8a8