Fix submitWholeTopic post-merge check
The fake Gerrit did not actually merge all of the changes that
should be submitted together simultaneously.  This meant our testing
that the git repo branch pointer moved forward was not correct.

This change improves the fake Gerrit so that all changes under
submitWholeTopic are merged simultaneously.

This then shows the error where after two changes are merged
simultaneously, the git repo check fails.

That is because our check is performed by simply ensuring that the
branch pointer is at some sha other than the one it was at right
before we started the merge operation.  The sequence is this:

1) Start reporting change #1
2) Get branch sha and store on change #1: abc1
3) Submit change #1  (Gerrit also merges change #2)
4) Get branch sha: abc2
5) Verify abc1 != abc2 (success)
6) Start reporting change #2
7) Get branch sha and store on change #2: abc2
8) Submit change #2  (already merged)
9) Get branch sha: abc2
A) Verify abc2 != abc2 (failure)

This is corrected by using only the first branch sha in a bundle.
When we store the branch sha on the change, we check to see if
any other changes in the bundle for the same project+branch already
have a sha; if they do, we use that.  Otherwise we know we are
the first, so we fetch it.

This changes the steps above to:

2) Get branch sha for all Gerrit changes in bundle
7) [not needed -- branch sha already stored in step 2]
A) Verify abc1 != abc2 (success)

Change-Id: Ia9ef411cbf24d1e4e31456ddc660e5b2a6eb5321