Only update commit dependencies when necessary
While I1c102b338ae6de05f3a752eb79f65744869e1860 fixed the referenced
exception it also introduced a performance problem. The issue was that
commit dependencies were always refreshed when a change was not found in
the local pipeline cache.

This cache is always empty when the pipeline's post-config hook is
called after a reconfiguration. Due to that this change could lead to
longer reconfigurations when a lot of items with dependencies were in
progress at that time.

Change-Id: I4d6c717903f17aa0770741ff948d9f594772891d