Handle non default loopvars in Ansible callback stream plugin
The Zuul Ansible callback stream plugin assumed that the ansible loop
var was always called 'item' in the result_dict. You can override this
value (and it is often necessary to do so to avoid collisions) to
something less generic. In those cases we would get errors like:

  b'[WARNING]: Failure using method (v2_runner_item_on_ok) in callback plugin'
  b'(<ansible.plugins.callback.zuul_stream.CallbackModule object at'
  b"0x7fbecc97c910>): 'item'"

And stream output would not include the info typically logged.

Address this by checking if ansible_loop_var is in the results_dict and
using that value for the loop var name instead. We still fall back to
'item' as I'm not sure that ansible_loop_var is always present.

Change-Id: I408e6d4af632f8097d63c04cbcb611d843086f6c